Resolved issues with deleting user and leagues

+++ src/main/java/com/kodilla/fantasy/service/UserDbService.java
@@ -1,15 +1,17 @@
 package com.kodilla.fantasy.service;
 
+import com.kodilla.fantasy.domain.League;
 import com.kodilla.fantasy.domain.Squad;
 import com.kodilla.fantasy.domain.User;
 import com.kodilla.fantasy.domain.exception.ElementNotFoundException;
 import com.kodilla.fantasy.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
+import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -27,16 +29,24 @@ public class UserDbService {
                 .orElseThrow(() -> new ElementNotFoundException("User with given id: " + id + " does not exist"));
     }
 
-    public Optional<User> getUserContainingSquad(Long id) {
-        return repository.findUserBySquadId(id);
-    }
-
     public User saveUser(User user) {
         return repository.save(user);
     }
 
-    public void deleteUser(Long id) {
-        repository.deleteById(id);
+    @Transactional
+    public void deleteUser(Long id) throws ElementNotFoundException {
+        if(repository.existsById(id)) {
+            repository.deleteById(id);
+        } else {
+            throw new ElementNotFoundException("User with given id: " + id + " does not exist");
+        }
+    }
+
+    public void detachUsers(League league) {
+        List<User> foundUsers = repository.findUserByLeaguesId(league.getId());
+        for(User user: foundUsers) {
+            user.getLeagues().remove(league);
+        }
     }
 
     @Transactional
